Backend/requirements.txt
# Minimal requirements for simple agent

livekit-agents[google]

python-dotenv

google-generativeai

supabase>=2.0.0

fastapi>=0.115.0

uvicorn>=0.32.0

